Remove some generic parameters.
Drastically simplify the code by making it not generic over the scalar type (stick to f64 as used in WebRender), and vector unit tags.
The unit tags don't play an important role in this crate since there is a single space. Letting the user of the crate cast to their own units makes the code a whole lot simpler.