Rewrite the BSP tree
The previous version of the BSP tree used a very naive data structure where each node is boxed and contains vectors of boxed children.
As a result any slightly complex tree results in large amounts of small memory allocations that can't be reused and very slow performance overall.

The new plane BSP tree does not change the logic, only how nodes polygons are stored using the classic vector with indices storage approach.

Some profiling of problematic test cases in Firefox showed that the BSP trees tend to be deep with usually few children per node. This commit takes advantage of that, using SmallVec in a few places.
In addition, profiling showed that cloning and moving the Polygon struct was occupying a significant portion if the profile, so polygons are now passed by reference in the hottest parts of the code.

This commit makes the plane splitting around 3 times faster on test cases such as the ones in https://bugzilla.mozilla.org/show_bug.cgi?id=1768984 and https://bugzilla.mozilla.org/show_bug.cgi?id=1493359.